multi: move PendingGroupWitness from asset to tapgarden
Relocates PendingGroupWitness to be closer to SealParams, where it is
actually used. Also moves UnmarshalGroupWitness to prevent an import
cycle.
